Getting ready for more architectures
Reorganise src/ from flat layout into fe/, ir/, amdgpu/ subdirectories.
Split amdgpu_emit.c (1800 lines) into emit.c, encode.c, enc_tab.c, and
encode.h to separate regalloc/asm/ELF from binary encoding. Zero include
changes — handled by -I flags in Makefile.
