+++ src/amdgpu/amdgpu.h
